requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
httpx
jinja2
google-generativeai
chromadb